astropy/1-SpectroscopicTraceTutorial/requirements.txt
astropy
astroquery>=0.4.8.dev9474
IPython
numpy
pillow
matplotlib